Move exchange offers list to separate component
