+++ src/screens/Exchange/Exchange.js
@@ -17,38 +17,30 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
+
 import * as React from 'react';
-import { FlatList, TextInput as RNTextInput, ScrollView, Keyboard } from 'react-native';
+import { TextInput as RNTextInput, ScrollView, Keyboard } from 'react-native';
 import type { NavigationEventSubscription, NavigationScreenProp } from 'react-navigation';
 import styled, { withTheme } from 'styled-components/native';
 import { connect } from 'react-redux';
 import debounce from 'lodash.debounce';
-import { formatAmount, formatMoney, formatFiat, isValidNumber, formatAmountDisplay } from 'utils/common';
+import { formatAmount, formatFiat, isValidNumber } from 'utils/common';
 import t from 'tcomb-form-native';
-import { utils } from 'ethers';
-import { BigNumber } from 'bignumber.js';
 import { createStructuredSelector } from 'reselect';
 import Intercom from 'react-native-intercom';
 import get from 'lodash.get';
 import isEmpty from 'lodash.isempty';
-import { InAppBrowser } from '@matt-block/react-native-in-app-browser';
 import { SDK_PROVIDER } from 'react-native-dotenv';
 
 // components
 import ContainerWithHeader from 'components/Layout/ContainerWithHeader';
 import TextInput from 'components/TextInput';
-import EmptyStateParagraph from 'components/EmptyState/EmptyStateParagraph';
 import SWActivationCard from 'components/SWActivationCard';
-import OfferCard from 'components/OfferCard/OfferCard';
 
 // actions
 import {
   searchOffersAction,
-  takeOfferAction,
-  authorizeWithShapeshiftAction,
   resetOffersAction,
-  setExecutingTransactionAction,
-  setTokenAllowanceAction,
   markNotificationAsSeenAction,
   getMetaDataAction,
   getExchangeSupportedAssetsAction,
@@ -56,20 +48,18 @@ import {
 import { hasSeenExchangeIntroAction } from 'actions/appSettingsActions';
 
 // constants
-import { EXCHANGE_CONFIRM, EXCHANGE_INFO, FIAT_EXCHANGE } from 'constants/navigationConstants';
+import { EXCHANGE_INFO } from 'constants/navigationConstants';
 import { defaultFiatCurrency, ETH, POPULAR_EXCHANGE_TOKENS, POPULAR_SWAPS } from 'constants/assetsConstants';
-import { PROVIDER_SHAPESHIFT } from 'constants/exchangeConstants';
 import { SMART_WALLET_UPGRADE_STATUSES } from 'constants/smartWalletConstants';
 import { ACCOUNT_TYPES } from 'constants/accountsConstants';
 
 // utils, services
-import { wyreWidgetUrl } from 'services/sendwyre';
 import { fiatCurrencies, initialAssets } from 'fixtures/assets';
 import { spacing } from 'utils/variables';
 import { getAssetData, getAssetsAsList, getBalance, getRate, sortAssets } from 'utils/assets';
-import { isFiatProvider, isFiatCurrency, getOfferProviderLogo } from 'utils/exchange';
+import { isFiatCurrency } from 'utils/exchange';
 import { getSmartWalletStatus } from 'utils/smartWallet';
-import { getActiveAccountType, getActiveAccountAddress } from 'utils/accounts';
+import { getActiveAccountType } from 'utils/accounts';
 import { themedColors } from 'utils/themes';
 import { satoshisToBtc } from 'utils/bitcoin';
 
@@ -79,7 +69,7 @@ import { paymentNetworkAccountBalancesSelector } from 'selectors/paymentNetwork'
 import { accountAssetsSelector } from 'selectors/assets';
 
 // models, types
-import type { Offer, FiatOffer, ExchangeSearchRequest, Allowance, ExchangeProvider, ProvidersMeta } from 'models/Offer';
+import type { ExchangeSearchRequest, Allowance, ExchangeProvider } from 'models/Offer';
 import type { Asset, Assets, Balances, Rates } from 'models/Asset';
 import type { SmartWalletStatus } from 'models/SmartWalletStatus';
 import type { Accounts } from 'models/Account';
@@ -88,52 +78,11 @@ import type { BitcoinAddress, BitcoinBalance } from 'models/Bitcoin';
 import type { Theme } from 'models/Theme';
 
 // partials
-import ExchangeStatus from './ExchangeStatus';
 import { HotSwapsHorizontalList } from './HotSwapsList';
 import ExchangeIntroModal from './ExchangeIntroModal';
+import ExchangeOffers from './ExchangeOffers';
+import { calculateMaxAmount, getFormattedBalanceInFiat } from './utils';
 
-const ListHeader = styled.View`
-  width: 100%;
-  align-items: flex-start;
-  margin-bottom: 8px;
-  padding: 0 ${spacing.layoutSides}px;
-`;
-
-const FormWrapper = styled.View`
-  padding: ${({ bottomPadding }) => `${spacing.large}px ${spacing.layoutSides}px ${bottomPadding}px`};
-  background-color: ${themedColors.surface};
-`;
-
-const ESWrapper = styled.View`
-  width: 100%;
-  align-items: center;
-  padding: 0 ${spacing.layoutSides}px;
-`;
-
-// const PromoWrapper = styled.View`
-//   width: 100%;
-//   align-items: center;
-//   padding: ${spacing.large}px ${spacing.layoutSides}px;
-//   margin-bottom: 30px;
-// `;
-//
-// const PromoText = styled(BaseText)`
-//   ${fontStyles.medium};
-//   color: ${themedColors.secondaryText};
-//   text-align: center;
-// `;
-//
-// const PopularSwapsGridWrapper = styled.View`
-//   border-top-width: 1px;
-//   border-bottom-width: 1px;
-//   border-color: ${themedColors.tertiary};
-//   background-color: ${themedColors.card};
-//   padding: ${spacing.large}px ${spacing.layoutSides}px 0;
-// `;
-
-const OfferCardWrapper = styled.View`
-  padding: 0 ${spacing.layoutSides}px;
-`;
 
 type Props = {
   rates: Rates,
@@ -142,19 +91,15 @@ type Props = {
   user: Object,
   assets: Assets,
   searchOffers: (string, string, number) => void,
-  offers: Offer[],
-  takeOffer: (string, string, number, string, string, Function) => void,
   authorizeWithShapeshift: Function,
   balances: Balances,
-  resetOffers: Function,
+  resetOffers: () => void,
   paymentNetworkBalances: Balances,
   exchangeSearchRequest: ExchangeSearchRequest,
-  setExecutingTransaction: Function,
-  setTokenAllowance: Function,
   exchangeAllowances: Allowance[],
   connectedProviders: ExchangeProvider[],
   hasUnreadExchangeNotification: boolean,
-  markNotificationAsSeen: Function,
+  markNotificationAsSeen: () => void,
   oAuthAccessToken: ?string,
   accounts: Accounts,
   smartWalletState: Object,
@@ -163,7 +108,6 @@ type Props = {
   exchangeSupportedAssets: Asset[],
   fiatExchangeSupportedAssets: Asset[],
   getExchangeSupportedAssets: () => void,
-  providersMeta: ProvidersMeta,
   hasSeenExchangeIntro: boolean,
   updateHasSeenExchangeIntro: () => void,
   btcAddresses: BitcoinAddress[],
@@ -171,73 +115,34 @@ type Props = {
   theme: Theme,
 };
 
+type InputValue = {
+  selector: Object,
+  input: string,
+}
+
+export type FormValue = {
+  fromInput: InputValue,
+  toInput: InputValue,
+}
+
 type State = {
-  value: Object,
-  shapeshiftAuthPressed: boolean,
+  value: FormValue,
   formOptions: Object,
-  // offer id will be passed to prevent double clicking
-  pressedOfferId: string,
-  pressedTokenAllowanceId: string,
   isSubmitted: boolean,
   showEmptyMessage: boolean,
 };
 
-const getAvailable = (_min, _max, rate) => {
-  if (!_min && !_max) {
-    return 'N/A';
-  }
-  let min = (new BigNumber(rate)).multipliedBy(_min);
-  let max = (new BigNumber(rate)).multipliedBy(_max);
-  if ((min.gte(0) && min.lt(0.01)) || (max.gte(0) && max.lt(0.01))) {
-    if (max.isZero()) return '>0.01';
-    const maxAvailable = max.lt(0.01)
-      ? '<0.01'
-      : formatMoney(max.toNumber(), 2);
-    return min.eq(max) || min.isZero()
-      // max available displayed if equal to min or min is zero
-      ? maxAvailable
-      : '<0.01 - <0.01';
-  }
-  min = min.toNumber();
-  max = max.toNumber();
-  if (!min || !max || min === max) {
-    return `${formatMoney(min || max, 2)}`;
-  }
-  return `${formatMoney(min, 2)} - ${formatMoney(max, 2)}`;
-};
+
+const FormWrapper = styled.View`
+  padding: ${({ bottomPadding }) => `${spacing.large}px ${spacing.layoutSides}px ${bottomPadding}px`};
+  background-color: ${themedColors.surface};
+`;
 
 const { Form } = t.form;
 
 const MIN_TX_AMOUNT = 0.000000000000000001;
 
 const settingsIcon = require('assets/icons/icon_key.png');
-
-const calculateMaxAmount = (token: string, balance: number | string): number => {
-  if (typeof balance !== 'string') {
-    balance = balance.toString();
-  }
-  if (token !== ETH) {
-    return +balance;
-  }
-  const maxAmount = utils.parseUnits(balance, 'ether');
-  if (maxAmount.lt(0)) return 0;
-  return new BigNumber(utils.formatEther(maxAmount)).toNumber();
-};
-
-const calculateAmountToBuy = (askRate: number | string, amountToSell: number | string) => {
-  return (new BigNumber(askRate)).multipliedBy(amountToSell).toFixed();
-};
-
-const getFormattedBalanceInFiat = (
-  baseFiatCurrency: ?string,
-  assetBalance: ?string | ?number,
-  rates: Rates,
-  symbol: string) => {
-  const fiatCurrency = baseFiatCurrency || defaultFiatCurrency;
-  const assetBalanceInFiat = assetBalance ?
-    parseFloat(assetBalance) * getRate(rates, symbol, fiatCurrency) : null;
-  return assetBalanceInFiat ? formatFiat(assetBalanceInFiat, fiatCurrency) : null;
-};
 
 const generateFormStructure = (balances: Balances) => {
   let balance;
@@ -384,53 +289,6 @@ function SelectorInputTemplate(locals) {
   );
 }
 
-function getCardAdditionalButtonData(additionalData) {
-  const {
-    offer,
-    minOrMaxNeeded,
-    isBelowMin,
-    thisComponent,
-    isShapeShift,
-    shapeshiftAccessToken,
-    storedAllowance,
-    allowanceSet,
-    shapeshiftAuthPressed,
-    pressedTokenAllowanceId,
-  } = additionalData;
-
-  const {
-    _id: offerId,
-    minQuantity,
-    maxQuantity,
-    fromAsset,
-  } = offer;
-
-  const { code: fromAssetCode } = fromAsset;
-  const isSetAllowancePressed = pressedTokenAllowanceId === offerId;
-  const minOrMaxAmount = formatAmountDisplay(isBelowMin ? minQuantity : maxQuantity);
-
-  if (minOrMaxNeeded) {
-    return {
-      title: `${isBelowMin ? 'Min' : 'Max'} ${minOrMaxAmount} ${fromAssetCode}`,
-      onPress: () => thisComponent.setFromAmount(isBelowMin ? minQuantity : maxQuantity),
-    };
-  } else if (isShapeShift && !shapeshiftAccessToken) {
-    return {
-      title: 'Connect',
-      onPress: thisComponent.onShapeshiftAuthPress,
-      disabled: shapeshiftAuthPressed,
-    };
-  } else if (!allowanceSet) {
-    return {
-      title: storedAllowance ? 'Pending' : 'Allow this exchange',
-      onPress: () => thisComponent.onSetTokenAllowancePress(offer),
-      disabled: isSetAllowancePressed,
-      isLoading: isSetAllowancePressed,
-    };
-  }
-  return null;
-}
-
 class ExchangeScreen extends React.Component<Props, State> {
   exchangeForm: t.form;
   fromInputRef: RNTextInput;
@@ -443,8 +301,6 @@ class ExchangeScreen extends React.Component<Props, State> {
     this.listeners = [];
     this.state = {
       shapeshiftAuthPressed: false,
-      pressedOfferId: '',
-      pressedTokenAllowanceId: '',
       isSubmitted: false,
       showEmptyMessage: false,
       value: {
@@ -658,6 +514,7 @@ class ExchangeScreen extends React.Component<Props, State> {
         const supportedAssetsOptions = this.generateSupportedAssetsOptions([toAsset]);
         initialFormState.toInput = {
           selector: supportedAssetsOptions[0],
+          input: '',
         };
       }
     }
@@ -701,140 +558,6 @@ class ExchangeScreen extends React.Component<Props, State> {
     return !(isFiatCurrency(from) && !fiatExchangeSupportedAssets.some(({ symbol }) => symbol === to));
   };
 
-  onShapeshiftAuthPress = () => {
-    const { authorizeWithShapeshift } = this.props;
-    this.setState({ shapeshiftAuthPressed: true }, async () => {
-      await authorizeWithShapeshift();
-      this.setState({ shapeshiftAuthPressed: false });
-    });
-  };
-
-  openSendWyre(selectedSellAmount: string, offer: FiatOffer) {
-    const { accounts, btcAddresses } = this.props;
-    const { fromAsset, toAsset } = offer;
-    const { code: fromAssetCode } = fromAsset;
-    const { code: toAssetCode } = toAsset;
-
-    let destAddress;
-    if (toAssetCode === 'BTC') {
-      destAddress = btcAddresses[0].address;
-    } else {
-      destAddress = getActiveAccountAddress(accounts);
-    }
-
-    const wyreUrl = wyreWidgetUrl(
-      destAddress,
-      toAssetCode,
-      fromAssetCode,
-      selectedSellAmount,
-    );
-
-    InAppBrowser.open(wyreUrl).catch(error => {
-      console.error('InAppBrowser.error', error); // eslint-disable-line no-console
-    });
-  }
-
-  onFiatOfferPress = (offer: FiatOffer) => {
-    const {
-      navigation,
-    } = this.props;
-    const {
-      value: {
-        fromInput: {
-          input: selectedSellAmount,
-        },
-      },
-    } = this.state;
-    const { provider } = offer;
-
-    if (provider === 'SendWyre') {
-      this.openSendWyre(selectedSellAmount, offer);
-      return;
-    }
-
-    navigation.navigate(FIAT_EXCHANGE, {
-      fiatOfferOrder: {
-        ...offer,
-        amount: selectedSellAmount,
-      },
-    });
-  };
-
-  onOfferPress = (offer: Offer) => {
-    const {
-      navigation,
-      takeOffer,
-      setExecutingTransaction,
-    } = this.props;
-    const {
-      value: {
-        fromInput: {
-          input: selectedSellAmount,
-        },
-      },
-    } = this.state;
-    const {
-      _id,
-      provider,
-      fromAsset,
-      toAsset,
-      askRate,
-      trackId = '',
-    } = offer;
-    const { code: fromAssetCode } = fromAsset;
-    const { code: toAssetCode } = toAsset;
-    const amountToSell = parseFloat(selectedSellAmount);
-    const amountToBuy = calculateAmountToBuy(askRate, amountToSell);
-    this.setState({ pressedOfferId: _id }, () => {
-      takeOffer(fromAssetCode, toAssetCode, amountToSell, provider, trackId, order => {
-        this.setState({ pressedOfferId: '' }); // reset offer card button loading spinner
-        if (isEmpty(order)) return;
-        setExecutingTransaction();
-        navigation.navigate(EXCHANGE_CONFIRM, {
-          offerOrder: {
-            ...order,
-            receiveQuantity: amountToBuy, // this value should be provided by exchange, currently returning 0,
-            // hence we overwrite it with our calculation
-            provider,
-          },
-        });
-      });
-    });
-  };
-
-  onSetTokenAllowancePress = (offer: Offer) => {
-    const {
-      navigation,
-      setTokenAllowance,
-      setExecutingTransaction,
-    } = this.props;
-    const {
-      _id,
-      provider,
-      fromAsset,
-      toAsset,
-      trackId = '',
-    } = offer;
-    const { address: fromAssetAddress, code: fromAssetCode } = fromAsset;
-    const { address: toAssetAddress } = toAsset;
-    this.setState({ pressedTokenAllowanceId: _id }, () => {
-      setTokenAllowance(fromAssetCode, fromAssetAddress, toAssetAddress, provider, trackId, (response) => {
-        this.setState({ pressedTokenAllowanceId: '' }); // reset set allowance button to be enabled
-        if (isEmpty(response)) return;
-        setExecutingTransaction();
-        navigation.navigate(EXCHANGE_CONFIRM, {
-          offerOrder: {
-            ...response,
-            provider,
-            fromAsset,
-            toAsset,
-            setTokenAllowance: true,
-          },
-        });
-      });
-    });
-  };
-
   setFromAmount = amount => {
     this.resetSearch(); // reset all cards before they change according to input values
     this.setState(prevState => ({
@@ -850,140 +573,6 @@ class ExchangeScreen extends React.Component<Props, State> {
       if (!isEmpty(errors)) return;
       this.triggerSearch();
     });
-  };
-
-  renderOffers = ({ item: offer }, disableNonFiatExchange: boolean) => {
-    const {
-      value: { fromInput },
-      pressedOfferId,
-      shapeshiftAuthPressed,
-      pressedTokenAllowanceId,
-    } = this.state;
-    const {
-      exchangeAllowances,
-      connectedProviders,
-      providersMeta,
-      theme,
-    } = this.props;
-    const { input: selectedSellAmount } = fromInput;
-    const {
-      _id: offerId,
-      minQuantity,
-      maxQuantity,
-      askRate,
-      toAsset,
-      fromAsset,
-      provider: offerProvider,
-      feeAmount,
-      extraFeeAmount,
-      quoteCurrencyAmount,
-      offerRestricted,
-    } = offer;
-    let { allowanceSet = true } = offer;
-
-    const { code: toAssetCode } = toAsset;
-    const { code: fromAssetCode } = fromAsset;
-
-    let storedAllowance;
-    if (!allowanceSet) {
-      storedAllowance = exchangeAllowances.find(
-        ({ provider, fromAssetCode: _fromAssetCode, toAssetCode: _toAssetCode }) => _fromAssetCode === fromAssetCode
-          && _toAssetCode === toAssetCode && provider === offerProvider,
-      );
-      allowanceSet = storedAllowance && storedAllowance.enabled;
-    }
-
-    const available = getAvailable(minQuantity, maxQuantity, askRate);
-    const amountToBuy = calculateAmountToBuy(askRate, selectedSellAmount);
-    const isTakeOfferPressed = pressedOfferId === offerId;
-    const isShapeShift = offerProvider === PROVIDER_SHAPESHIFT;
-    const providerLogo = getOfferProviderLogo(providersMeta, offerProvider, theme, 'horizontal');
-    const amountToBuyString = formatAmountDisplay(amountToBuy);
-
-    let shapeshiftAccessToken;
-    if (isShapeShift) {
-      ({ extra: shapeshiftAccessToken } = connectedProviders
-        .find(({ id: providerId }) => providerId === PROVIDER_SHAPESHIFT) || {});
-    }
-
-    const amountToSell = parseFloat(selectedSellAmount);
-    const minQuantityNumeric = parseFloat(minQuantity);
-    const maxQuantityNumeric = parseFloat(maxQuantity);
-    const isBelowMin = minQuantityNumeric !== 0 && amountToSell < minQuantityNumeric;
-    const isAboveMax = maxQuantityNumeric !== 0 && amountToSell > maxQuantityNumeric;
-
-    const minOrMaxNeeded = isBelowMin || isAboveMax;
-    const isTakeButtonDisabled = !!minOrMaxNeeded
-      || isTakeOfferPressed
-      || !allowanceSet
-      || (isShapeShift && !shapeshiftAccessToken);
-
-    const isFiat = isFiatProvider(offerProvider);
-
-    const disableFiatExchange = isFiat && (minOrMaxNeeded || !!offerRestricted);
-    const disableOffer = disableNonFiatExchange || disableFiatExchange;
-
-    const additionalData = {
-      offer,
-      minOrMaxNeeded,
-      isBelowMin,
-      isShapeShift,
-      shapeshiftAccessToken,
-      allowanceSet,
-      storedAllowance,
-      shapeshiftAuthPressed,
-      pressedTokenAllowanceId,
-      thisComponent: this,
-    };
-
-    if (isFiat) {
-      return (
-        <OfferCardWrapper>
-          <OfferCard
-            isDisabled={isTakeButtonDisabled || disableOffer}
-            onPress={() => this.onFiatOfferPress(offer)}
-            labelTop="Amount total"
-            valueTop={`${askRate} ${fromAssetCode}`}
-            cardImageSource={providerLogo}
-            labelBottom="Fees total"
-            valueBottom={feeAmount ?
-              `${formatAmountDisplay(feeAmount + extraFeeAmount)} ${fromAssetCode}`
-              : 'Will be calculated'
-            }
-            cardButton={{
-              title: `${formatAmountDisplay(quoteCurrencyAmount)} ${toAssetCode}`,
-              onPress: () => this.onFiatOfferPress(offer),
-              disabled: disableFiatExchange,
-              isLoading: isTakeOfferPressed,
-            }}
-
-            cardNote={offerRestricted}
-            additionalCardButton={getCardAdditionalButtonData(additionalData)}
-          />
-        </OfferCardWrapper>
-      );
-    }
-
-    return (
-      <OfferCardWrapper>
-        <OfferCard
-          isDisabled={isTakeButtonDisabled || disableOffer}
-          onPress={() => this.onOfferPress(offer)}
-          labelTop="Exchange rate"
-          valueTop={formatAmountDisplay(askRate)}
-          cardImageSource={providerLogo}
-          labelBottom="Available"
-          valueBottom={available}
-          cardButton={{
-            title: `${amountToBuyString} ${toAssetCode}`,
-            onPress: () => this.onOfferPress(offer),
-            disabled: isTakeButtonDisabled || disableNonFiatExchange,
-            isLoading: isTakeOfferPressed,
-          }}
-          additionalCardButton={getCardAdditionalButtonData(additionalData)}
-        />
-      </OfferCardWrapper>
-    );
   };
 
   generateAssetsOptions = (assets: Assets) => {
@@ -1158,7 +747,6 @@ class ExchangeScreen extends React.Component<Props, State> {
 
   render() {
     const {
-      offers,
       balances,
       navigation,
       exchangeAllowances,
@@ -1182,7 +770,6 @@ class ExchangeScreen extends React.Component<Props, State> {
     const { selector: selectedToOption } = toInput;
 
     const formStructure = generateFormStructure(balances);
-    const reorderedOffers = offers.sort((a, b) => (new BigNumber(b.askRate)).minus(a.askRate).toNumber());
     const rightItems = [{ label: 'Support', onPress: () => Intercom.displayMessenger(), key: 'getHelp' }];
     if ((!isEmpty(exchangeAllowances) || !isEmpty(connectedProviders))
       && !rightItems.find(({ key }) => key === 'exchangeSettings')) {
@@ -1207,11 +794,7 @@ class ExchangeScreen extends React.Component<Props, State> {
     const isSelectedFiat = !isEmpty(selectedFromOption) &&
       fiatCurrencies.some(({ symbol }) => symbol === selectedFromOption.symbol);
     const disableNonFiatExchange = !this.checkIfAssetsExchangeIsAllowed() && !isSelectedFiat;
-    const flatListContentStyle = {
-      width: '100%',
-      paddingVertical: 10,
-      flexGrow: 1,
-    };
+
 
     const swaps = this.generatePopularSwaps();
 
@@ -1257,43 +840,15 @@ class ExchangeScreen extends React.Component<Props, State> {
               buttonTitle="Activate Smart Wallet"
             />
           }
-          {!!isSubmitted && (
-            <FlatList
-              data={reorderedOffers}
-              keyExtractor={(item) => item._id}
-              style={{ width: '100%', flex: 1 }}
-              keyboardShouldPersistTaps="handled"
-              contentContainerStyle={flatListContentStyle}
-              renderItem={(props) => this.renderOffers(props, disableNonFiatExchange)}
-              ListHeaderComponent={(
-                <ListHeader>
-                  <ExchangeStatus isVisible={isSubmitted && isSupportedExchange} />
-                </ListHeader>
-              )}
-              ListEmptyComponent={!!showEmptyMessage && (
-                <ESWrapper style={{ marginTop: '15%', marginBottom: spacing.large }}>
-                  <EmptyStateParagraph
-                    title="No live offers"
-                    bodyText="Currently no matching offers from exchange services are provided.
-                              New offers may appear at any time — don’t miss it."
-                    large
-                    wide
-                  />
-                </ESWrapper>
-              )}
-              // ListFooterComponentStyle={{ flex: 1, justifyContent: 'flex-end' }}
-              // ListFooterComponent={
-              //   <PopularSwapsGridWrapper>
-              //     <SafeAreaView forceInset={{ top: 'never', bottom: 'always' }}>
-              //       <MediumText medium style={{ marginBottom: spacing.medium }}>
-              //           Try these popular swaps
-              //       </MediumText>
-              //       <HotSwapsGridList onPress={this.onSwapPress} swaps={swaps} />
-              //     </SafeAreaView>
-              //   </PopularSwapsGridWrapper>
-              // }
-            />
-          )}
+          {!!isSubmitted &&
+          <ExchangeOffers
+            value={value}
+            disableNonFiatExchange={disableNonFiatExchange}
+            isExchangeActive={isSubmitted && isSupportedExchange}
+            showEmptyMessage={showEmptyMessage}
+            setFromAmount={this.setFromAmount}
+            navigation={navigation}
+          />}
         </ScrollView>}
       </ContainerWithHeader>
     );
@@ -1305,13 +860,11 @@ const mapStateToProps = ({
   appSettings: { data: { baseFiatCurrency, hasSeenExchangeIntro } },
   exchange: {
     data: {
-      offers,
       searchRequest: exchangeSearchRequest,
       allowances: exchangeAllowances,
       connectedProviders,
       hasNotification: hasUnreadExchangeNotification,
     },
-    providersMeta,
     exchangeSupportedAssets,
     fiatExchangeSupportedAssets,
   },
@@ -1331,7 +884,6 @@ const mapStateToProps = ({
   },
 }: RootReducerState): $Shape<Props> => ({
   baseFiatCurrency,
-  offers,
   rates,
   exchangeSearchRequest,
   exchangeAllowances,
@@ -1341,7 +893,6 @@ const mapStateToProps = ({
   smartWalletFeatureEnabled,
   accounts,
   smartWalletState,
-  providersMeta,
   exchangeSupportedAssets,
   fiatExchangeSupportedAssets,
   hasSeenExchangeIntro,
@@ -1364,15 +915,7 @@ const mapDispatchToProps = (dispatch: Dispatch): $Shape<Props> => ({
   searchOffers: (fromAssetCode, toAssetCode, fromAmount) => dispatch(
     searchOffersAction(fromAssetCode, toAssetCode, fromAmount),
   ),
-  takeOffer: (fromAssetCode, toAssetCode, fromAmount, provider, trackId, callback) => dispatch(
-    takeOfferAction(fromAssetCode, toAssetCode, fromAmount, provider, trackId, callback),
-  ),
-  authorizeWithShapeshift: () => dispatch(authorizeWithShapeshiftAction()),
   resetOffers: () => dispatch(resetOffersAction()),
-  setExecutingTransaction: () => dispatch(setExecutingTransactionAction()),
-  setTokenAllowance: (formAssetCode, fromAssetAddress, toAssetAddress, provider, trackId, callback) => dispatch(
-    setTokenAllowanceAction(formAssetCode, fromAssetAddress, toAssetAddress, provider, trackId, callback),
-  ),
   markNotificationAsSeen: () => dispatch(markNotificationAsSeenAction()),
   getMetaData: () => dispatch(getMetaDataAction()),
   getExchangeSupportedAssets: () => dispatch(getExchangeSupportedAssetsAction()),
